AC-652: Pulled grant check into dashboard instead of routes
