+++ src/pages/dashboard/DashboardPage.js
@@ -1,7 +1,4 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
-import _ from 'lodash';
 
 import { Page } from '@sparkpost/matchbox';
 import { UsageReport } from 'src/components';
@@ -9,26 +6,17 @@ import Tutorial from './components/Tutorial';
 import VerifyEmailBanner from 'src/components/verifyEmailBanner/VerifyEmailBanner';
 import SuppressionBanner from './components/SuppressionBanner';
 import { FreePlanWarningBanner } from 'src/pages/billing/components/Banners';
-
-import { fetch as fetchAccount } from 'src/actions/account';
-import { checkSuppression } from 'src/actions/suppressions';
-import { list as listSendingDomains } from 'src/actions/sendingDomains';
-import { listApiKeys } from 'src/actions/api-keys';
-
-import { selectAccountAgeInWeeks , selectAccountAgeInDays } from 'src/selectors/accountAge';
-import { selectVerifiedDomains, selectNotBlockedDomains } from 'src/selectors/sendingDomains';
-import { selectApiKeysForSending } from 'src/selectors/api-keys';
-import { hasGrants } from 'src/helpers/conditions';
-
 export class DashboardPage extends Component {
   componentDidMount() {
-    this.props.checkSuppression();
-    this.props.listSendingDomains();
-    this.props.listApiKeys({ id: 0 });
+    if (this.props.canViewTutorialAndSuppressions) {
+      this.props.checkSuppression();
+      this.props.listSendingDomains();
+      this.props.listApiKeys({ id: 0 });
+    }
   }
 
   render() {
-    const { accountAgeInWeeks, accountAgeInDays, currentUser, hasSuppressions, account } = this.props;
+    const { accountAgeInWeeks, accountAgeInDays, currentUser, hasSuppressions, account, canViewTutorialAndSuppressions } = this.props;
 
     //Shows banner if within 14 days of plan to downgrade
 
@@ -39,32 +27,15 @@ export class DashboardPage extends Component {
         )}
         <FreePlanWarningBanner account={account} accountAgeInDays={accountAgeInDays} ageRangeStart={16}/>
         <UsageReport accountAgeInWeeks={accountAgeInWeeks} />
-        <SuppressionBanner accountAgeInWeeks={accountAgeInWeeks} hasSuppressions={hasSuppressions} />
-        <Tutorial {...this.props} />
+        {canViewTutorialAndSuppressions && (
+          <>
+            <SuppressionBanner accountAgeInWeeks={accountAgeInWeeks} hasSuppressions={hasSuppressions} />
+            <Tutorial {...this.props} />
+          </>
+        )}
       </Page>
     );
   }
 }
 
-function mapStateToProps(state) {
-  const acctAgeWeeks = selectAccountAgeInWeeks(state);
-  const acctAgeDays = selectAccountAgeInDays(state);
-  const notBlockedDomains = selectNotBlockedDomains(state);
-  const verifiedDomains = selectVerifiedDomains(state);
-  const apiKeysForSending = selectApiKeysForSending(state);
-
-  return {
-    account: state.account,
-    currentUser: state.currentUser,
-    accountAgeInWeeks: acctAgeWeeks,
-    accountAgeInDays: acctAgeDays,
-    hasSuppressions: state.suppressions.hasSuppression,
-    hasSendingDomains: notBlockedDomains.length > 0,
-    hasVerifiedDomains: verifiedDomains.length > 0,
-    hasApiKeysForSending: apiKeysForSending.length > 0,
-    hasSentThisMonth: _.get(state, 'account.usage.month.used', 0) > 0,
-    canViewTutorial: hasGrants('api_keys/manage', 'templates/modify', 'sending_domains/manage')
-  };
-}
-
-export default withRouter(connect(mapStateToProps, { fetchAccount, checkSuppression, listSendingDomains, listApiKeys })(DashboardPage));
+export default DashboardPage;
