+++ src/pages/dashboard/index.js
@@ -1,0 +1,3 @@
+import DashboardPage from './DashboardPage.container';
+
+export default DashboardPage;
